README CHALLENGE 1
CHALLENGE 1 (EASY):
Inspect Element - not a save way to store your password.

In this challenge you must get the login credentials trough inspecting element. In the HTML code there is a hidden encrypted script which stores the username and password in it.
Users need to search for a decrypt tool to decrypt the encrypt script. The hint button shows examples of decrypt tools.

When the users decrypt the script, A username and password combination can be found which are the same. 

When the user found the username and password then they can try to fill in the login form on the website. When correct a popup will tell the user he has completed challenge 1 and
he may continue.